Use `\S*` instead of `.*` to capture VolumeTag
Currently, VolumeTags are captured with a `.*`, which also captures any whitespaces appearing after the Tag. It's causing issues when gathering that tags because the spaces are part of the tag for this package, whereas they are just polluting the tag.

Using `\S*` would stop the capture as soon as it reach a whitespace character. I don't expect Tags to contain spaces.